Starting to use new redirect uri config param.
This is mostly for reference and future development. Basic idea is to use the
"redirect-uri" param (which needs to be set in Azure/AD/Applications in ms) to
drive all urls. This may be local or it may be some reverse proxy to handle ssl.

Also starting to move SSL to be optional on the local server since, as stated
above it may be handled elsewhere.